move assistant functions to own file, add input format options, validations
