+++ VISOR/VISOR.py
@@ -164,7 +164,7 @@ def main():
 	required.add_argument('-s','--sample', help='one folder containing FASTA haplotypes such as those generated by HACk', metavar='FOLDER', required=True)
 	required.add_argument('-b','--bedfile', help='one or more regions to simulate in standard BED format', metavar='BED', required=True)
 	required.add_argument('-o','--output', help='output folder with for FASTQ files', metavar='FOLDER', required=True)
-	required.add_argument('-O','--output-format', help='output format for FASTQ files [10x] {choices: 10x, stlfr, haplotagging, tellseq}', metavar='', choices = ["10x", "stlfr", "haplotagging", "tellseq"])
+	required.add_argument('-O','--output-format', help='output format for FASTQ files [10x] {choices: 10x, stlfr, haplotagging, tellseq}', default="10x", metavar='', choices = ["10x", "stlfr", "haplotagging", "tellseq"])
 
 	wgsim = parser_linkedreads.add_argument_group('Paired-end FASTQ simulation using pywgsim')
 
@@ -180,6 +180,7 @@ def main():
 	molecules=parser_linkedreads.add_argument_group('Linked-read simulation')
 	bc_default = os.path.abspath(os.path.dirname(__file__) + '/XENIA/4M-with-alts-february-2016.txt.gz')
 	molecules.add_argument('--barcodes', help='file of barcodes to simulate, one per line [10X default barcodes]', metavar='FILE', default = bc_default, type=barcodefile)
+	molecules.add_argument('--barcode-type', help='Type of linked-read barcodes used for --barcodes [10x] {choices: 10x, stlfr, haplotagging, tellseq}', metavar='', default = "10x", choices = ["10x", "stlfr", "haplotagging", "tellseq"])
 	molecules.add_argument('--molecule_length', help='mean length of molecules [80000]', metavar='', default=80000, type=int)
 	molecules.add_argument('--molecule_number', help='number of fragments per barcode on average [10]', metavar='', default=10, type=int)
 	molecules.add_argument('--molecule_coverage', help='mean coverage per molecule [0.2]', metavar='', default=0.2, type=float)
